Moved utils to correct folder struct
