+++ ros2_ws/src/python_package/python_package/inference_node.py
